requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.10.0
blis==0.7.7
bs4==0.0.1
catalogue==2.0.7
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cymem==2.0.6
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
filetype==1.0.10
Flask==2.0.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
googletrans==3.1.0a0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.1
langcodes==3.3.0
langdetect==1.0.9
MarkupSafe==2.1.1
murmurhash==1.0.6
numpy==1.22.3
packaging==21.3
pathy==0.6.1
preshed==3.0.6
puremagic==1.12
pydantic==1.8.2
pyparsing==3.0.7
python-magic==0.4.25
python-magic-bin==0.4.14
pytz==2022.1
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.1
spacy==3.2.3
spacy-language-detection==0.2.1
spacy-legacy==3.0.9
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.15
tqdm==4.63.1
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.9
wasabi==0.9.0
Werkzeug==2.0.3
